feat(auth,resume): improve auth forms and resume builder UI
This commit enhances the user experience for authentication and the resume builder.

- Improves error handling and feedback on sign-in and sign-up forms.
- Refactors the education, skills, and personal info sections for better UI/UX.
- Extracts the personal info form into its own component.
- Adds an empty state for the personal info section.
- Makes trusted origins more flexible for local development.
