+++ src/features/resume-builder/sections/skills-section.tsx
@@ -12,7 +12,6 @@ import { Progress } from '@/shared/components/ui/progress';
 import { AnimatePresenceLazy } from '@/shared/utilities/dynamic-motion';
 import { addSkillCategory, removeSkillCategory, updateSkillCategory } from '@/store/resume-store';
 import type { TSkillCategory } from '@/types/resume';
-import { FormSection } from '../form/form-section';
 
 function importSkillsForm() {
 	return import('./skills-form').then(function map(mod) {
@@ -103,7 +102,19 @@ export function SkillsSection({ data }: TProps) {
 	}
 
 	return (
-		<FormSection title='Skills' icon={<Code className='h-5 w-5' />}>
+		<div className='max-w-4xl mx-auto p-8 space-y-8'>
+			<div className='flex items-center gap-3 mb-6'>
+				<div className='w-10 h-10 bg-primary/20 rounded-lg flex items-center justify-center'>
+					<Code className='h-5 w-5 text-primary' />
+				</div>
+				<div>
+					<h1 className='text-2xl font-semibold text-foreground'>Skills</h1>
+					<p className='text-muted-foreground'>
+						Highlight your technical and professional expertise
+					</p>
+				</div>
+			</div>
+
 			<div className='space-y-6'>
 				<AnimatePresenceLazy>
 					{isFormVisible && (
@@ -192,6 +203,6 @@ export function SkillsSection({ data }: TProps) {
 					</div>
 				)}
 			</div>
-		</FormSection>
+		</div>
 	);
 }
